Cambio masivo en cuanto comandos y el comando help

+++ bot/comandos/moderacion/deletec.js
@@ -1,7 +1,15 @@
-module.exports = (client, message, args) => {
-  if (!message.member.permissions.has("ADMINISTRATOR")) return message.channel.send("No puedes usar ese comando")
+module.exports = {
+  name: "deletec",
+  description: "Elimina el canal en el que te encuentras",
+  use: "",
+  category: 'moderacion',
+  alias: [],
+  async run(client, message, args) {
+    const Discord = require("discord.js")
+  if (!message.member.permissions.has("ADMINISTRATOR") || !process.env.OWNERS_ID.includes(message.author.id)) return message.channel.send("No puedes usar ese comando")
   message.channel.send("Adios :)").then(() => {
     message.channel.delete()
   })
   
+}
 }
